+++ RELEASE_NOTES.md
@@ -1,13 +1,22 @@
 # Changelog
 
+## 1.0.5
+
+- Updated decimal using mpdec++ source
+- Fixed memory leak while converting bigdecimal to string
+- Updated gtest to 10
+
 ## 1.0.4
+
 - Automatically increase precision (if neccessary) while using operators `*=`, `+=` and `/=`
 
 ## 1.0.3
+
 - Updated mpdecimal conan dependency to 2.5.1
 - Fixed `_MSC_VER` redefinition
 
 ## 1.0.2 (Fri Jan 22 2021)
+
 - Updated mpdecimal conan dependency to 2.5.0 (as 2.4.2 no more compiles on macOS Big Sur)
 
 ## 1.0.1 (Thu Jan 14 2021)